refactor: move repository interfaces to domain layer and update imports
